requirements.txt

torch
numpy
open3d
opencv-python
pyglet=1.5.0
pyyaml
tqdm
trimesh
